feat(sonarr): download client (#63)
* feat(sonarr): add downloadclientconfig

* feat(sonarr): add downloadclient

* feat(sonarr): add remotepathmapping

* fix: align comments fix typos and path joining

* style: forgot a path join

* fix: int64 type and few more typo

* fix: wrong assertion in test

+++ sonarr/languageprofile.go
@@ -46,12 +46,12 @@ func (s *Sonarr) GetLanguageProfilesContext(ctx context.Context) ([]*LanguagePro
 }
 
 // GetLanguageProfile returns a single language profile.
-func (s *Sonarr) GetLanguageProfile(profileID int) (*LanguageProfile, error) {
+func (s *Sonarr) GetLanguageProfile(profileID int64) (*LanguageProfile, error) {
 	return s.GetLanguageProfileContext(context.Background(), profileID)
 }
 
 // GetLanguageProfileContext returns a single language profile.
-func (s *Sonarr) GetLanguageProfileContext(ctx context.Context, profileID int) (*LanguageProfile, error) {
+func (s *Sonarr) GetLanguageProfileContext(ctx context.Context, profileID int64) (*LanguageProfile, error) {
 	var output LanguageProfile
 
 	req := starr.Request{URI: path.Join(bpLanguageProfile, fmt.Sprint(profileID))}
@@ -98,7 +98,7 @@ func (s *Sonarr) UpdateLanguageProfileContext(ctx context.Context, profile *Lang
 		return nil, fmt.Errorf("json.Marshal(%s): %w", bpLanguageProfile, err)
 	}
 
-	req := starr.Request{URI: path.Join(bpLanguageProfile, fmt.Sprint(int(profile.ID))), Body: &body}
+	req := starr.Request{URI: path.Join(bpLanguageProfile, fmt.Sprint(profile.ID)), Body: &body}
 	if err := s.PutInto(ctx, req, &output); err != nil {
 		return nil, fmt.Errorf("api.Put(%s): %w", &req, err)
 	}
@@ -107,12 +107,12 @@ func (s *Sonarr) UpdateLanguageProfileContext(ctx context.Context, profile *Lang
 }
 
 // DeleteLanguageProfile removes a single language profile.
-func (s *Sonarr) DeleteLanguageProfile(profileID int) error {
+func (s *Sonarr) DeleteLanguageProfile(profileID int64) error {
 	return s.DeleteLanguageProfileContext(context.Background(), profileID)
 }
 
 // DeleteLanguageProfileContext removes a single language profile.
-func (s *Sonarr) DeleteLanguageProfileContext(ctx context.Context, profileID int) error {
+func (s *Sonarr) DeleteLanguageProfileContext(ctx context.Context, profileID int64) error {
 	req := starr.Request{URI: path.Join(bpLanguageProfile, fmt.Sprint(profileID))}
 	if err := s.DeleteAny(ctx, req); err != nil {
 		return fmt.Errorf("api.Delete(%s): %w", &req, err)
