+++ sonarr/rootfolder.go
@@ -40,12 +40,12 @@ func (s *Sonarr) GetRootFoldersContext(ctx context.Context) ([]*RootFolder, erro
 }
 
 // GetRootFolder returns a single root folder.
-func (s *Sonarr) GetRootFolder(folderID int) (*RootFolder, error) {
+func (s *Sonarr) GetRootFolder(folderID int64) (*RootFolder, error) {
 	return s.GetRootFolderContext(context.Background(), folderID)
 }
 
 // GetRootFolderContext returns a single root folder.
-func (s *Sonarr) GetRootFolderContext(ctx context.Context, folderID int) (*RootFolder, error) {
+func (s *Sonarr) GetRootFolderContext(ctx context.Context, folderID int64) (*RootFolder, error) {
 	var output RootFolder
 
 	req := starr.Request{URI: path.Join(bpRootFolder, fmt.Sprint(folderID))}
@@ -79,12 +79,12 @@ func (s *Sonarr) AddRootFolderContext(ctx context.Context, folder *RootFolder) (
 }
 
 // DeleteRootFolder removes a single root folder.
-func (s *Sonarr) DeleteRootFolder(folderID int) error {
+func (s *Sonarr) DeleteRootFolder(folderID int64) error {
 	return s.DeleteRootFolderContext(context.Background(), folderID)
 }
 
 // DeleteRootFolderContext removes a single root folder.
-func (s *Sonarr) DeleteRootFolderContext(ctx context.Context, folderID int) error {
+func (s *Sonarr) DeleteRootFolderContext(ctx context.Context, folderID int64) error {
 	req := starr.Request{URI: path.Join(bpRootFolder, fmt.Sprint(folderID))}
 	if err := s.DeleteAny(ctx, req); err != nil {
 		return fmt.Errorf("api.Delete(%s): %w", &req, err)
